Add editing and deleting settings

+++ Views/SetupMenu.axaml.cs
@@ -2,6 +2,7 @@
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Interactivity;
+using Avalonia.Media;
 using Avalonia.Threading;
 using MedEye.DB;
 
@@ -12,7 +13,9 @@ public partial class SetupMenu : Window
     private static readonly DispatcherTimer CloseTimer = new DispatcherTimer();
 
     private int _currentGame = 0;
-    private List<Settings> _games;
+    private int _userId = -1;
+    private int _gameNumber = -1;
+    private List<Settings> _gamesSettings;
     private DispatcherTimer NextGameTimer = new DispatcherTimer();
 
 
@@ -23,11 +26,20 @@ public partial class SetupMenu : Window
         this.AttachDevTools();
 #endif
 
-        MainMenu.Click += MainMenuClick;
+        MainMenu.Click += (s, e) =>
+        {
+            new ConfirmAction(this.ClientSize.Width, this.ClientSize.Height,
+                "Вы уверены что хотите выйти в главное меню? Настройки приложения не будут сохранены!",
+                MainMenuClick).Show();
+        };
         StartGame.Click += StartGameClick;
         AddGame.Click += AddGameHandle;
         SaveGame.Click += SaveGameHandle;
-        DeleteGame.Click += DeleteGameHandle;
+        DeleteGame.Click += (s, e) =>
+        {
+            new ConfirmAction(this.ClientSize.Width, this.ClientSize.Height,
+                "Вы уверены, что хотите удалить игру из списка?", DeleteGameHandle).Show();
+        };
 
         GamesScroll.Height = ClientSize.Height / 1.2;
         GamesScroll.MaxHeight = ClientSize.Height / 1.2;
@@ -36,42 +48,73 @@ public partial class SetupMenu : Window
         CloseTimer.Tick += CloseAfterRoute;
         CloseTimer.Interval = new TimeSpan(1000000);
 
-        foreach (var settings in SettingsWrap.GetSettings(0))
+        _gamesSettings = SettingsWrap.GetSettings(_userId);
+        foreach (var settings in _gamesSettings)
         {
             SettingsWrap.DeleteSettings(settings);
+            _gamesSettings.Remove(settings);
         }
     }
 
-    public SetupMenu(int user_id)
+    public SetupMenu(int userId)
     {
         InitializeComponent();
 #if DEBUG
         this.AttachDevTools();
 #endif
 
-        MainMenu.Click += MainMenuClick;
+        _userId = userId;
+
+        MainMenu.Click += (s, e) =>
+        {
+            new ConfirmAction(this.ClientSize.Width, this.ClientSize.Height,
+                "Вы уверены, что хотите перейти в главное меню? Все несохраненные измения будут потеряны!",
+                MainMenuClick).Show();
+        };
         StartGame.Click += StartGameClick;
         AddGame.Click += AddGameHandle;
         SaveGame.Click += SaveGameHandle;
-        DeleteGame.Click += DeleteGameHandle;
+        DeleteGame.Click += (s, e) =>
+        {
+            new ConfirmAction(this.ClientSize.Width, this.ClientSize.Height,
+                "Вы уверены, что хотите удалить игру?", DeleteGameHandle).Show();
+        };
 
         GamesScroll.Height = ClientSize.Height / 1.2;
         GamesScroll.MaxHeight = ClientSize.Height / 1.2;
 
+        Games.Children.Remove(DefaultGame);
 
         CloseTimer.Tick += CloseAfterRoute;
         CloseTimer.Interval = new TimeSpan(1000000);
 
-        foreach (var settings in SettingsWrap.GetSettings(0))
+        _gamesSettings = SettingsWrap.GetSettings(_userId);
+        if (_userId == -1)
         {
-            SettingsWrap.DeleteSettings(settings);
+            foreach (var settings in _gamesSettings)
+            {
+                SettingsWrap.DeleteSettings(settings);
+                _gamesSettings.Remove(settings);
+            }
+        }
+        else
+        {
+            foreach (var settings in _gamesSettings)
+            {
+                AddGameToGrid();
+                _gameNumber++;
+                var game = (ComboBox)Games.Children[_gameNumber];
+                game.SelectedIndex = settings.GameId - 1;
+            }
+
+            ((ComboBox)Games.Children[_gameNumber]).FontWeight = FontWeight.Bold;
+            SetSettings(_gamesSettings[_gameNumber]);
         }
     }
 
     protected override void OnOpened(EventArgs e)
     {
         AdaptToScreen();
-
         base.OnOpened(e);
     }
 
@@ -83,6 +126,17 @@ public partial class SetupMenu : Window
 
     private void CloseAfterRoute(object? sender, EventArgs e)
     {
+        foreach (var setting in SettingsWrap.GetSettings(_userId))
+        {
+            SettingsWrap.DeleteSettings(setting);
+        }
+
+        var priority = 1;
+        foreach (var setting in _gamesSettings)
+        {
+            SettingsWrap.AddSettings(setting.SetPriority(priority++));
+        }
+
         Close();
         CloseTimer.Stop();
     }
@@ -97,7 +151,7 @@ public partial class SetupMenu : Window
             .OfType<RadioButton>()
             .FirstOrDefault(r => r.IsChecked ?? false);
 
-        var gameId = ((ComboBox)Games.Children[^1]).SelectedIndex + 1;
+        var gameId = ((ComboBox)Games.Children[_gameNumber]).SelectedIndex + 1;
         var priority = Games.Children.Count;
         var distanceValue = distance is null ? 0 : int.Parse((string)distance.Content);
         var isRedValue = isRed is not null && (isRed.IsChecked ?? false);
@@ -111,7 +165,7 @@ public partial class SetupMenu : Window
 
         var settings = new Settings
         {
-            UserId = 0,
+            UserId = _userId,
             GameId = gameId,
             Priority = priority,
             Distance = distanceValue,
@@ -147,15 +201,41 @@ public partial class SetupMenu : Window
         TimerTextBox.Text = "5";
     }
 
+    private void SetSettings(Settings settings)
+    {
+        var distance = DistanceRadioButtons.Children
+            .OfType<RadioButton>()
+            .FirstOrDefault(r => int.Parse((string)r.Content) == settings.Distance);
+        if (distance != null) distance.IsChecked = true;
+
+        if (settings.IsRed)
+        {
+            var isRed = IsRedRadioButtons.Children.OfType<RadioButton>().First();
+            isRed.IsChecked = true;
+        }
+        else
+        {
+            var isNotRed = IsRedRadioButtons.Children.OfType<RadioButton>().Last();
+            isNotRed.IsChecked = true;
+        }
+
+        FrequencyFlickerBox.SelectedIndex = settings.Frequency / 10 - 1;
+        TypeFlickerBox.SelectedIndex = settings.FlickerMode;
+        BrightRedColorSlider.Value = settings.RedBrightness;
+        BrightBlueColorSlider.Value = settings.BlueBrightness;
+        LevelSlider.Value = settings.Level;
+        TimerTextBox.Text = (settings.ExerciseDuration / 60.0).ToString(CultureInfo.InvariantCulture);
+    }
+
     private void NextGame(object? sender, EventArgs e)
     {
         NextGameTimer.Stop();
 
         _currentGame++;
 
-        if (_currentGame >= _games.Count) return;
+        if (_currentGame >= _gamesSettings.Count) return;
 
-        var game = _games[_currentGame];
+        var game = _gamesSettings[_currentGame];
         switch (game.GameId)
         {
             case 1:
@@ -170,6 +250,8 @@ public partial class SetupMenu : Window
             case 4:
                 new Merger(game).Show();
                 break;
+            default:
+                return;
         }
 
         NextGameTimer.Interval = new TimeSpan(0, 0, game.ExerciseDuration + 5);
@@ -180,12 +262,11 @@ public partial class SetupMenu : Window
     {
         NextGameTimer.Stop();
 
-        _games = SettingsWrap.GetSettings(0);
         _currentGame = 0;
-        
-        if (_games.Count == 0) return;
-        
-        var game = _games[_currentGame];
+
+        if (_gamesSettings.Count == 0) return;
+
+        var game = _gamesSettings[_currentGame];
         switch (game.GameId)
         {
             case 1:
@@ -200,6 +281,8 @@ public partial class SetupMenu : Window
             case 4:
                 new Merger(game).Show();
                 break;
+            default:
+                return;
         }
 
         NextGameTimer.Tick += NextGame;
@@ -209,25 +292,41 @@ public partial class SetupMenu : Window
 
     private void SaveGameHandle(object? sender, EventArgs e)
     {
-        var settings = GetCurrentSettings();
+        if (_gameNumber == -1) return;
 
-        SettingsWrap.AddSettings(settings);
+        var settings = GetCurrentSettings();
+        if (_gameNumber < _gamesSettings.Count)
+        {
+            _gamesSettings[_gameNumber] = settings;
+        }
+        else
+        {
+            _gamesSettings.Insert(_gameNumber, settings);
+        }
     }
-    
+
     private void DeleteGameHandle(object? sender, EventArgs e)
     {
         if (Games.Children.Count == 0) return;
-        
-        var game = Games.Children[^1];
+
+        var game = Games.Children[_gameNumber];
+        _gamesSettings.RemoveAt(_gameNumber);
+
+        _gameNumber -= 1;
+        if (_gameNumber > -1)
+        {
+            ((ComboBox)Games.Children[_gameNumber]).FontWeight = FontWeight.Bold;
+            SetSettings(_gamesSettings[_gameNumber]);
+        }
 
         Games.Children.Remove(game);
     }
 
-    private void AddGameHandle(object? sender, RoutedEventArgs e)
+    private void AddGameToGrid()
     {
         var game = new ComboBox
         {
-            PlaceholderText = Game1.PlaceholderText,
+            PlaceholderText = DefaultGame.PlaceholderText,
             Items = new ComboBoxItem[]
             {
                 new() { Content = "Тир" },
@@ -239,20 +338,63 @@ public partial class SetupMenu : Window
             FontSize = 32 * (ClientSize.Width / 1920)
         };
 
-        Games.RowDefinitions.Add(new RowDefinition());
+        game.Tapped += ClickGame;
+        if (Games.RowDefinitions.Count == Games.Children.Count)
+        {
+            Games.RowDefinitions.Add(new RowDefinition());
+        }
 
         Grid.SetRow(game, Games.Children.Count);
         Games.Children.Add(game);
+    }
 
+    private void AddGameHandle(object? sender, RoutedEventArgs e)
+    {
+        if (_gameNumber > -1)
+        {
+            ((ComboBox)Games.Children[_gameNumber]).FontWeight = FontWeight.Normal;
+        }
+
+        _gameNumber = Games.Children.Count;
+        AddGameToGrid();
+        ((ComboBox)Games.Children[_gameNumber]).FontWeight = FontWeight.Bold;
         ResetSettings();
+    }
+
+    private void ClickGame(object? sender, RoutedEventArgs e)
+    {
+        if (sender == null) return;
+        var newGame = (ComboBox)sender;
+        var oldGame = (ComboBox)Games.Children[_gameNumber];
+
+        if (newGame.Equals(oldGame)) return;
+
+        newGame.FontWeight = FontWeight.Bold;
+        oldGame.FontWeight = FontWeight.Normal;
+
+        _gameNumber = Games.Children.IndexOf(newGame);
+        if (_gameNumber < _gamesSettings.Count)
+        {
+            SetSettings(_gamesSettings[_gameNumber]);
+        }
+        else
+        {
+            ResetSettings();
+        }
     }
 
     private void AdaptToScreen()
     {
         var buttonWidth = 2 * this.ClientSize.Width / 9;
+        var fontSize = 32 * (this.ClientSize.Width / 1920);
 
+        foreach (var comboBox in Games.Children.OfType<ComboBox>())
+        {
+            comboBox.Width = buttonWidth;
+            comboBox.FontSize = fontSize;
+        }
 
-        Game1.Width = buttonWidth;
+        DefaultGame.Width = buttonWidth;
         MainMenu.Width = buttonWidth;
         DeleteGame.Width = buttonWidth;
         StartGame.Width = buttonWidth;
@@ -261,22 +403,22 @@ public partial class SetupMenu : Window
         AddGame.Width = buttonWidth;
         SaveGame.Width = buttonWidth;
 
-        Game1.FontSize = 32 * (this.ClientSize.Width / 1920);
-        MainMenu.FontSize = 32 * (this.ClientSize.Width / 1920);
-        DeleteGame.FontSize = 32 * (this.ClientSize.Width / 1920);
-        StartGame.FontSize = 32 * (this.ClientSize.Width / 1920);
-        Distance.FontSize = 32 * (this.ClientSize.Width / 1920);
-        LeftFilterColor.FontSize = 32 * (this.ClientSize.Width / 1920);
-        FrequencyFlickerText.FontSize = 32 * (this.ClientSize.Width / 1920);
-        FrequencyFlickerBox.FontSize = 32 * (this.ClientSize.Width / 1920);
-        TypeFlickerText.FontSize = 32 * (this.ClientSize.Width / 1920);
-        TypeFlickerBox.FontSize = 32 * (this.ClientSize.Width / 1920);
-        BrightRedColor.FontSize = 32 * (this.ClientSize.Width / 1920);
-        BrightBlueColor.FontSize = 32 * (this.ClientSize.Width / 1920);
-        Level.FontSize = 32 * (this.ClientSize.Width / 1920);
-        Timer.FontSize = 32 * (this.ClientSize.Width / 1920);
-        AddGame.FontSize = 32 * (this.ClientSize.Width / 1920);
-        SaveGame.FontSize = 32 * (this.ClientSize.Width / 1920);
+        DefaultGame.FontSize = fontSize;
+        MainMenu.FontSize = fontSize;
+        DeleteGame.FontSize = fontSize;
+        StartGame.FontSize = fontSize;
+        Distance.FontSize = fontSize;
+        LeftFilterColor.FontSize = fontSize;
+        FrequencyFlickerText.FontSize = fontSize;
+        FrequencyFlickerBox.FontSize = fontSize;
+        TypeFlickerText.FontSize = fontSize;
+        TypeFlickerBox.FontSize = fontSize;
+        BrightRedColor.FontSize = fontSize;
+        BrightBlueColor.FontSize = fontSize;
+        Level.FontSize = fontSize;
+        Timer.FontSize = fontSize;
+        AddGame.FontSize = fontSize;
+        SaveGame.FontSize = fontSize;
 
         Header1.FontSize = 48 * (this.ClientSize.Width / 1920);
         Header2.FontSize = 48 * (this.ClientSize.Width / 1920);
